+++ CHANGELOG.md
@@ -5,17 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- Update prop `handleTabChange` in `Tabbed` component;
+
+- Add new prop `inputClassName` in `Select` component.
+
 ## [0.39.0] - 2021-07-07
 
 - Update Dropdown:
         
-    - Add new props: `onDropdownVisibleChange`, `autoFocus`.
+    - Add new props: `onDropdownVisibleChange`, `autoFocus`;
 
-- Add new component `Select`
+- Add new component `Select`;
 
 - Update Tabbed component:
     
-    - Add new prop `size`    
+    - Add new prop `size`.
     
 ## [0.38.0] - 2021-06-28